+++ src/Oro/Bundle/PayPalBundle/Migrations/Schema/OroPayPalBundleInstaller.php
@@ -16,7 +16,7 @@ class OroPayPalBundleInstaller implements Installation
      */
     public function getMigrationVersion()
     {
-        return 'v1_1';
+        return 'v1_2';
     }
 
     /**
@@ -46,18 +46,42 @@ class OroPayPalBundleInstaller implements Installation
         $table->addColumn('pp_credit_card_action', 'string', ['notnull' => false, 'length' => 255]);
         $table->addColumn('pp_allowed_card_types', 'array', ['notnull' => false, 'comment' => '(DC2Type:array)']);
         $table->addColumn('pp_express_checkout_name', 'string', ['notnull' => false, 'length' => 255]);
-        $table->addColumn('pp_partner', 'string', ['notnull' => false, 'length' => 255]);
-        $table->addColumn('pp_vendor', 'string', ['notnull' => false, 'length' => 255]);
-        $table->addColumn('pp_user', 'string', ['notnull' => false, 'length' => 255]);
-        $table->addColumn('pp_password', 'string', ['notnull' => false, 'length' => 255]);
+        $table->addColumn('pp_partner', 'string', [
+            'notnull' => false,
+            'length' => 255,
+            'comment' => '(DC2Type:crypted_string)'
+        ]);
+        $table->addColumn('pp_vendor', 'crypted_string', [
+            'notnull' => false,
+            'length' => 255,
+            'comment' => '(DC2Type:crypted_string)'
+        ]);
+        $table->addColumn('pp_user', 'crypted_string', [
+            'notnull' => false,
+            'length' => 255,
+            'comment' => '(DC2Type:crypted_string)'
+        ]);
+        $table->addColumn('pp_password', 'crypted_string', [
+            'notnull' => false,
+            'length' => 255,
+            'comment' => '(DC2Type:crypted_string)'
+        ]);
         $table->addColumn('pp_test_mode', 'boolean', ['default' => '0', 'notnull' => false]);
         $table->addColumn('pp_debug_mode', 'boolean', ['default' => '0', 'notnull' => false]);
         $table->addColumn('pp_require_cvv_entry', 'boolean', ['default' => '1', 'notnull' => false]);
         $table->addColumn('pp_zero_amount_authorization', 'boolean', ['default' => '0', 'notnull' => false]);
         $table->addColumn('pp_auth_for_req_amount', 'boolean', ['default' => '0', 'notnull' => false]);
         $table->addColumn('pp_use_proxy', 'boolean', ['default' => '0', 'notnull' => false]);
-        $table->addColumn('pp_proxy_host', 'string', ['notnull' => false, 'length' => 255]);
-        $table->addColumn('pp_proxy_port', 'string', ['notnull' => false, 'length' => 255]);
+        $table->addColumn('pp_proxy_host', 'crypted_string', [
+            'notnull' => false,
+            'length' => 255,
+            'comment' => '(DC2Type:crypted_string)'
+        ]);
+        $table->addColumn('pp_proxy_port', 'crypted_string', [
+            'notnull' => false,
+            'length' => 255,
+            'comment' => '(DC2Type:crypted_string)'
+        ]);
         $table->addColumn('pp_enable_ssl_verification', 'boolean', ['default' => '1', 'notnull' => false]);
     }