+++ src/Oro/Bundle/PayPalBundle/Tests/Unit/Form/Type/PayPalSettingsTypeTest.php
@@ -2,14 +2,12 @@
 
 namespace Oro\Bundle\PayPalBundle\Tests\Unit\Form\Type;
 
-use Oro\Bundle\FormBundle\Form\Type\OroEncodedPlaceholderPasswordType;
 use Oro\Bundle\LocaleBundle\Form\Type\LocalizedFallbackValueCollectionType;
 use Oro\Bundle\LocaleBundle\Tests\Unit\Form\Type\Stub\LocalizedFallbackValueCollectionTypeStub;
 use Oro\Bundle\PayPalBundle\Entity\PayPalSettings;
 use Oro\Bundle\PayPalBundle\Form\Type\PayPalSettingsType;
 use Oro\Bundle\PayPalBundle\Settings\DataProvider\CreditCardTypesDataProviderInterface;
 use Oro\Bundle\PayPalBundle\Settings\DataProvider\PaymentActionsDataProviderInterface;
-use Oro\Bundle\SecurityBundle\Encoder\SymmetricCrypterInterface;
 use Oro\Component\Testing\Unit\FormIntegrationTestCase;
 use Oro\Component\Testing\Unit\PreloadedExtension;
 use Symfony\Component\Form\Extension\Validator\ValidatorExtension;
@@ -31,11 +29,6 @@ class PayPalSettingsTypeTest extends FormIntegrationTestCase
      * @var PayPalSettingsType
      */
     private $formType;
-
-    /**
-     * @var SymmetricCrypterInterface|\PHPUnit\Framework\MockObject\MockObject
-     */
-    private $encoder;
 
     /**
      * @var TranslatorInterface|\PHPUnit\Framework\MockObject\MockObject
@@ -71,10 +64,8 @@ class PayPalSettingsTypeTest extends FormIntegrationTestCase
                 'charge',
             ]);
 
-        $this->encoder = $this->createMock(SymmetricCrypterInterface::class);
         $this->formType = new PayPalSettingsType(
             $this->translator,
-            $this->encoder,
             $creditCardTypesDataProvider,
             $this->paymentActionsDataProvider
         );
@@ -94,7 +85,6 @@ class PayPalSettingsTypeTest extends FormIntegrationTestCase
                 [
                     PayPalSettingsType::class => $this->formType,
                     LocalizedFallbackValueCollectionType::class => $localizedType,
-                    OroEncodedPlaceholderPasswordType::class => new OroEncodedPlaceholderPasswordType($this->encoder),
                 ],
                 []
             ),
@@ -132,18 +122,6 @@ class PayPalSettingsTypeTest extends FormIntegrationTestCase
             'proxyPort' => 'port',
             'enableSSLVerification' => false,
         ];
-
-        $this->encoder
-            ->expects(static::any())
-            ->method('encryptData')
-            ->willReturnMap([
-                [$submitData['vendor'], $submitData['vendor']],
-                [$submitData['partner'], $submitData['partner']],
-                [$submitData['user'], $submitData['user']],
-                [$submitData['password'], $submitData['password']],
-                [$submitData['proxyHost'], $submitData['proxyHost']],
-                [$submitData['proxyPort'], $submitData['proxyPort']],
-            ]);
 
         $payPalSettings = new PayPalSettings();
 
@@ -211,7 +189,6 @@ class PayPalSettingsTypeTest extends FormIntegrationTestCase
 
         $formType = new PayPalSettingsType(
             $this->translator,
-            $this->encoder,
             $creditCardTypesDataProvider,
             $this->paymentActionsDataProvider
         );