+++ src/Oro/Bundle/PayPalBundle/Tests/Unit/OptionsProvider/LineItemOptionsFormatterTest.php
@@ -1,54 +1,88 @@
 <?php
 
-namespace Oro\Bundle\PayPalBundle\Tests\Unit\EventListener;
+namespace Oro\Bundle\PayPalBundle\Tests\Unit\OptionsProvider;
 
 use Oro\Bundle\CurrencyBundle\Rounding\RoundingServiceInterface;
 use Oro\Bundle\LocaleBundle\Formatter\NumberFormatter;
-use Oro\Bundle\OrderBundle\Entity\Order;
-use Oro\Bundle\PaymentBundle\Event\ExtractLineItemPaymentOptionsEvent;
 use Oro\Bundle\PaymentBundle\Model\LineItemOptionModel;
-use Oro\Bundle\PayPalBundle\EventListener\ExtractLineItemPaymentOptionsListener;
+use Oro\Bundle\PayPalBundle\OptionsProvider\LineItemOptionsFormatter;
+use PHPUnit\Framework\MockObject\MockObject;
+use PHPUnit\Framework\TestCase;
 
-class ExtractLineItemPaymentOptionsListenerTest extends \PHPUnit\Framework\TestCase
+class LineItemOptionsFormatterTest extends TestCase
 {
-    /** @var ExtractLineItemPaymentOptionsListener */
-    private $listener;
+    /** @var NumberFormatter|MockObject */
+    private $numberFormatter;
 
-    /** @var RoundingServiceInterface|\PHPUnit\Framework\MockObject\MockObject */
+    /** @var RoundingServiceInterface|MockObject */
     private $rounder;
 
-    public function setUp()
+    /** @var LineItemOptionsFormatter */
+    private $lineItemOptionsFormatter;
+
+    protected function setUp()
     {
-        /** @var NumberFormatter|\PHPUnit\Framework\MockObject\MockObject $numberFormatter */
-        $numberFormatter = $this->getMockBuilder(NumberFormatter::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-        $numberFormatter->expects($this->any())
+        $this->numberFormatter = $this->createMock(NumberFormatter::class);
+        $this->rounder = $this->createMock(RoundingServiceInterface::class);
+        $this->lineItemOptionsFormatter = new LineItemOptionsFormatter($this->numberFormatter, $this->rounder);
+    }
+
+    /**
+     * @param array $lineItemModelData
+     * @param int $currencyRoundingPrecision
+     * @param array $expected
+     * @dataProvider formatLineItemOptionsDataProvider
+     */
+    public function testFormatLineItemOptions(
+        array $lineItemModelData,
+        int $currencyRoundingPrecision,
+        array $expected
+    ): void {
+        $lineItemModel = new LineItemOptionModel();
+        $lineItemModel->setName($lineItemModelData['name']);
+        $lineItemModel->setDescription($lineItemModelData['description']);
+        $lineItemModel->setCost($lineItemModelData['cost']);
+        $lineItemModel->setQty($lineItemModelData['qty']);
+        $lineItemModel->setCurrency($lineItemModelData['currency']);
+        $lineItemModel->setUnit($lineItemModelData['unit']);
+
+        $this->numberFormatter
+            ->expects($this->any())
             ->method('formatCurrency')
             ->willReturnCallback(function ($cost, $currency) {
                 return sprintf('%s%s', $currency, round($cost, 2));
             });
 
-        $this->rounder = $this->getMockBuilder(RoundingServiceInterface::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-        $this->rounder->expects($this->any())
+        $this->rounder
+            ->expects($this->any())
             ->method('round')
             ->willReturnCallback(function ($number, $precision) {
                 return round($number, $precision);
             });
 
-        $this->listener = new ExtractLineItemPaymentOptionsListener($numberFormatter, $this->rounder);
+        $this->rounder->expects($this->any())
+            ->method('getPrecision')
+            ->willReturn($currencyRoundingPrecision);
+
+        $formattedLineItemOptions = $this->lineItemOptionsFormatter->formatLineItemOptions([$lineItemModel]);
+        $actual = reset($formattedLineItemOptions);
+
+        $this->assertEquals($expected['name'], $actual->getName());
+        $this->assertEquals($expected['description'], $actual->getDescription());
+        $this->assertEquals($expected['cost'], round($actual->getCost(), 2), '', 1e-6);
+        $this->assertEquals($expected['qty'], $actual->getQty(), '', 1e-6);
+        $this->assertEquals($expected['currency'], $actual->getCurrency());
+        $this->assertEquals($expected['unit'], $actual->getUnit());
     }
 
     /**
      * @return array
      * @SuppressWarnings(PHPMD.ExcessiveMethodLength)
      */
-    public function onExtractLineItemPaymentOptionsData()
+    public function formatLineItemOptionsDataProvider(): array
     {
         return [
-            'Whole qty, cost precision 2' => [
+            'Integer qty, cost precision 2' => [
                 'lineItemModelData' => [
                     'name' => 'PRSKU Product Name',
                     'description' => 'Product Description',
@@ -86,7 +120,7 @@ class ExtractLineItemPaymentOptionsListenerTest extends \PHPUnit\Framework\TestC
                     'unit' => 'kg',
                 ],
             ],
-            'Whole qty, cost precision 3' => [
+            'Integer qty, cost precision 3' => [
                 'lineItemModelData' => [
                     'name' => 'PRSKU Long Product Name',
                     'description' => 'Product Description',
@@ -95,7 +129,7 @@ class ExtractLineItemPaymentOptionsListenerTest extends \PHPUnit\Framework\TestC
                     'currency' => 'EUR',
                     'unit' => 'kg',
                 ],
-                'currencyRoundingPrecision' => 2,
+                'currencyRoundingPrecision' => 3,
                 'expected' => [
                     'name' => 'PRSKU Long Product N - EUR16.67x2 kg',
                     'description' => 'Product Description',
@@ -114,7 +148,7 @@ class ExtractLineItemPaymentOptionsListenerTest extends \PHPUnit\Framework\TestC
                     'currency' => 'EUR',
                     'unit' => 'kg',
                 ],
-                'currencyRoundingPrecision' => 2,
+                'currencyRoundingPrecision' => 3,
                 'expected' => [
                     'name' => 'PRSKU Long Product - EUR13.34x0.2 kg',
                     'description' => 'Product Description',
@@ -144,47 +178,5 @@ class ExtractLineItemPaymentOptionsListenerTest extends \PHPUnit\Framework\TestC
                 ],
             ],
         ];
-    }
-
-    /**
-     * @param array $lineItemModelData
-     * @param int $currencyRoundingPrecision
-     * @param array $expected
-     * @dataProvider onExtractLineItemPaymentOptionsData
-     */
-    public function testOnExtractLineItemPaymentOptions(
-        array $lineItemModelData,
-        $currencyRoundingPrecision,
-        array $expected
-    ) {
-        $this->rounder->expects($this->any())
-            ->method('getPrecision')
-            ->willReturn($currencyRoundingPrecision);
-
-        $lineItemModel = new LineItemOptionModel();
-
-        $lineItemModel->setName($lineItemModelData['name']);
-        $lineItemModel->setDescription($lineItemModelData['description']);
-        $lineItemModel->setCost($lineItemModelData['cost']);
-        $lineItemModel->setQty($lineItemModelData['qty']);
-        $lineItemModel->setCurrency($lineItemModelData['currency']);
-        $lineItemModel->setUnit($lineItemModelData['unit']);
-
-        $entity = new Order();
-        $event = new ExtractLineItemPaymentOptionsEvent($entity);
-
-        $this->assertEmpty($event->getModels());
-        $event->addModel($lineItemModel);
-
-        $this->listener->onExtractLineItemPaymentOptions($event);
-
-        $result = $event->getModels()[0];
-
-        $this->assertEquals($expected['name'], $result->getName());
-        $this->assertEquals($expected['description'], $result->getDescription());
-        $this->assertEquals($expected['cost'], round($result->getCost(), 2), '', 1e-6);
-        $this->assertEquals($expected['qty'], $result->getQty(), '', 1e-6);
-        $this->assertEquals($expected['currency'], $result->getCurrency());
-        $this->assertEquals($expected['unit'], $result->getUnit());
     }
 }